# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduQuest Gamified Education Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthRequest {
  username: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  xp: number;
  level: number;
  levelName: string;
  streak: number;
  badges: string[];
  xpToNextLevel: number;
  progressPercent: number;
  completedMissions: number[];
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface DailyLoginResponse {
  xpAwarded: number;
  streakBonus: number;
  newStreak: number;
  message: string;
  alreadyClaimed: boolean;
}

export type MissionType = (typeof MissionType)[keyof typeof MissionType];

export const MissionType = {
  quiz: "quiz",
  lesson: "lesson",
  project: "project",
} as const;

export type MissionDifficulty =
  (typeof MissionDifficulty)[keyof typeof MissionDifficulty];

export const MissionDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Mission {
  id: number;
  title: string;
  description: string;
  type: MissionType;
  xpReward: number;
  completed: boolean;
  category: string;
  difficulty: MissionDifficulty;
}

export interface MissionCompleteResponse {
  xpAwarded: number;
  newXp: number;
  newLevel: number;
  levelName: string;
  leveledUp: boolean;
  badgesUnlocked: string[];
  message: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  xp: number;
  level: number;
  levelName: string;
  badges: string[];
}

export interface CourseRecommendation {
  title: string;
  description: string;
  difficulty: string;
  estimatedXp: number;
  tags: string[];
}

export interface RecommendationResponse {
  recommended: string;
  reason: string;
  nextCourses: CourseRecommendation[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}