# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduQuest Gamified Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  username: string;
  xp: number;
  level: number;
  levelName: string;
  streak: number;
  badges: string[];
  xpToNextLevel: number;
  progressPercent: number;
  completedMissions: number[];
}